+++ pkgs/test_core/lib/src/runner/reporter/compact.dart
@@ -10,7 +10,6 @@ import 'package:test_api/src/backend/live_test.dart'; // ignore: implementation_
 import 'package:test_api/src/backend/message.dart'; // ignore: implementation_imports
 import 'package:test_api/src/backend/state.dart'; // ignore: implementation_imports
 import 'package:test_api/src/utils.dart'; // ignore: implementation_imports
-// ignore: implementation_imports
 import 'package:test_api/src/utils.dart' as utils;
 
 import '../../util/io.dart';
@@ -253,6 +252,12 @@ class CompactReporter implements Reporter {
       if (!_printedNewline) stdout.write(" " * (lineLength - message.length));
       stdout.writeln();
     } else if (!success) {
+      for (var liveTest in _engine.active) {
+        _progressLine(_description(liveTest),
+            truncate: false,
+            suffix: " - did not complete $_bold$_red[E]$_noColor");
+        print('');
+      }
       _progressLine('Some tests failed.', color: _red);
       print('');
     } else if (_engine.passed.isEmpty) {